LayoutTests/http/tests/security/mixedContent/insecure-audio-video-in-main-frame-UpgradeMixedContent.html

<html><!-- webkit-test-runner [ UpgradeMixedContentEnabled=true ] -->
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

window.addEventListener("message", function (e) {
    if (window.testRunner)
        testRunner.notifyDone();
}, false);

</script>
<p>This test opens a window that loads HTML5 audio and video after upgrading
the URL's scheme from http: to https:.  We should trigger a mixed content
callback because the main frame in the window is HTTPS but the requested resource was HTTP.</p>
<script>
onload = function() {
    window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-audio-video.html");
}
</script>
</body>
</html>
