LayoutTests/http/tests/security/mixedContent/insecure-audio-video-in-main-frame-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-audio-video.html requested insecure content from http://127.0.0.1:8080/resources/test.mp4. This content was automatically upgraded and should be served over HTTPS.

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-audio-video.html requested insecure content from http://127.0.0.1:8080/resources/test.mp4. This content was automatically upgraded and should be served over HTTPS.

This test opens a window that loads HTML5 audio and video after upgrading the URL's scheme from http: to https:. We should trigger a mixed content callback because the main frame in the window is HTTPS but the requested resource was HTTP.
